% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-irods.R
\name{create_irods}
\alias{create_irods}
\title{Generate IRODS project file}
\usage{
create_irods(host, path = ".", overwrite = FALSE)
}
\arguments{
\item{host}{URL of host.}

\item{path}{Path of project (defaults to current directory).}

\item{overwrite}{Overwrite existing irods file (defaults to \code{FALSE}).}
}
\value{
Invisibly returns path to irods file.
}
\description{
This will create an irods project file containing information about the
iRODS server.
}
\examples{

# connect project to server
create_irods("http://localhost/irods-rest/0.9.2")
}
