% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{get_token}
\alias{get_token}
\title{Authentication service for the iRODS zone}
\usage{
get_token(host = "http://localhost/irods-rest/0.9.2")
}
\arguments{
\item{host}{Url of host}
}
\value{
String
}
\description{
Authentication service for the iRODS zone
}
\examples{

# get token
get_token()

}
