% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{icd}
\alias{icd}
\alias{ipwd}
\title{Get or set current working directory in iRODS}
\usage{
icd(dir)

ipwd()
}
\arguments{
\item{dir}{Collection to set as working directory.}
}
\value{
Invisibly the current directory before the change (same convention as
\code{setwd()}).
}
\description{
\code{ipwd()} and \code{icd()} are the iRODS equivalents of \code{getwd()} and \code{setwd()}
respectively. For example, whereas \code{getwd()} returns the current working directory
in the local system, \code{ipwd()} returns the current working directory in iRODS.
}
\examples{
# demonstration server (requires Bash, Docker and Docker-compose)
irods_demo <- try(use_irods_demo())

if (!inherits(irods_demo, "try-error")) {

  # move to temporary directory and save old working directory
  old_dir <- getwd()
  tmp <- tempdir()
  setwd(tmp)

  # connect project to server
  create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home",
    overwrite = TRUE)

  # authenticate
  iauth("rods", "rods")

  # default dir
  icd(".")
  ipwd()

  # relative paths work as well
  icd("/tempZone/home")
  ipwd()

  # go back on level lower
  icd("..")
  ipwd()

  # absolute paths work as well
  icd("/tempZone/home/rods")
  ipwd()

  stop_irods_demo()

  # back to previous directory
  setwd(old_dir)
}
}
