% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-objects.R
\name{ichksum}
\alias{ichksum}
\title{Calculate or Retrieve a Data Object Checksum in iRODS}
\usage{
ichksum(
  logical_path,
  resource = NULL,
  replica_number = NULL,
  force = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{logical_path}{Path to the data object in iRODS.}

\item{resource}{Optional resource name identifying the replica to checksum.}

\item{replica_number}{Optional replica number identifying the replica to
checksum.}

\item{force}{Whether any existing checksum should be recalculated. Defaults
to \code{FALSE}.}

\item{verbose}{Whether information should be printed about the HTTP request
and response. Defaults to \code{FALSE}.}
}
\value{
A checksum string.
}
\description{
Request the checksum for a data object from the iRODS server using the
HTTP API \code{calculate_checksum} operation.
}
\examples{
if (is_irods_demo_running()) {
  \dontshow{
  .old_config_dir <- Sys.getenv("R_USER_CONFIG_DIR")
  .old_wd <- setwd(tempdir())
  Sys.setenv("R_USER_CONFIG_DIR" = tempdir())
  }
  # connect project to server
  \Sexpr[stage=build, results=rd]{paste0("create_irods(\"", rirods:::.irods_host, "\")")}

  # authenticate
  iauth("rods", "rods")

  # save a file and request its checksum
  write.csv(iris, "iris.csv")
  iput("iris.csv", "iris.csv")
  ichksum("iris.csv")

  # clean up
  irm("iris.csv", force = TRUE)
  \dontshow{
  setwd(.old_wd)
  Sys.setenv("R_USER_CONFIG_DIR" = .old_config_dir)
  }
}
}
\seealso{
\code{\link{iput()}} for storing files and \code{\link{iget()}} for retrieving them.
}
