% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{ils}
\alias{ils}
\title{List iRODS data objects and collections}
\usage{
ils(
  logical_path = ".",
  stat = FALSE,
  permissions = FALSE,
  metadata = FALSE,
  offset = 0,
  limit = 100,
  message = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{logical_path}{Path to the collection whose contents are to be listed.
By default this is the current working directory (see \code{ipwd()}).}

\item{stat}{Whether stat information should be included. Defaults to \code{FALSE}.}

\item{permissions}{Whether access control information should be included.
Defaults to \code{FALSE}.}

\item{metadata}{Whether metadata information should be included. Defaults to
\code{FALSE}.}

\item{offset}{Number of records to skip for pagination. Defaults to 0.}

\item{limit}{Number of records to show per page. Defaults to 100.}

\item{message}{Whether a message should be printed when the collection is
empty. Defaults to \code{TRUE}.}

\item{verbose}{Whether information should be printed about the HTTP request
and response. Defaults to \code{FALSE}.}
}
\value{
Dataframe with logical paths and, if requested, additional information.
}
\description{
List the contents of a collection, optionally with stat, metadata, and/or
access control information for each element in the collection.
}
\examples{
# demonstration server (requires Bash, Docker and Docker-compose)
irods_demo <- try(use_irods_demo())

if (!inherits(irods_demo, "try-error")) {

  # move to temporary directory and save old working directory
  old_dir <- getwd()
  tmp <- tempdir()
  setwd(tmp)

  # connect project to server
  create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home",
    overwrite = TRUE)

  # authenticate
  iauth("rods", "rods")

  # list home directory
  ils()

  # make collection
  imkdir("some_collection")

  # list a different directory
  ils("/tempZone/home/rods/some_collection")

  # show metadata
  ils(metadata = TRUE)

  stop_irods_demo()

  # back to previous directory
  setwd(old_dir)
}
}
