% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{imeta}
\alias{imeta}
\title{Title}
\usage{
imeta(
  x,
  entity_type,
  operations = list(operation = "add", attribute = NULL, value = NULL, units = NULL),
  path = ".",
  verbose = FALSE
)
}
\arguments{
\item{x}{object, collection or user}

\item{entity_type}{Type (object, collection or user)}

\item{operations}{List which contains the following named elements;
\code{operation}, which adds (\code{"add"}) or removes (\code{"remove"}) the object, and
the metadata triplet attribute, value, units.}

\item{path}{Path to object (defaults to \code{"."}).}

\item{verbose}{Show information about the http request and response.}
}
\value{
Invisibly returns the response.
}
\description{
Title
}
