% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{imkdir}
\alias{imkdir}
\title{Create a new collection in iRODS}
\usage{
imkdir(logical_path, create_parent_collections = FALSE, verbose = FALSE)
}
\arguments{
\item{logical_path}{Path to the collection to create, relative to the current
working directory (see \code{ipwd()}).}

\item{create_parent_collections}{Whether parent collections should be created
when necessary. Defaults to \code{FALSE}.}

\item{verbose}{Whether information about the HTTP request and response
should be printed. Defaults to \code{FALSE}.}
}
\value{
Invisibly the HTTP request.
}
\description{
This is the equivalent to \code{dir.create()}, but creating a collection in iRODS
instead of a local directory.
}
\examples{
# demonstration server (requires Bash, Docker and Docker-compose)
irods_demo <- try(use_irods_demo())

if (!inherits(irods_demo, "try-error")) {

  # move to temporary directory and save old working directory
  old_dir <- getwd()
  tmp <- tempdir()
  setwd(tmp)

  # connect project to server
  create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home", overwrite = TRUE)

  # authentication
  iauth("rods", "rods")

  # list all object and collection in the current collection of iRODS
  ils()

  # create a new collection
  imkdir("new_collection")

  # check if it is there
  ils()

  # and move to the new directory
  icd("new_collection")

  stop_irods_demo()

  # back to previous directory
  setwd(old_dir)
}
}
