% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{iquery}
\alias{iquery}
\title{Query data objects and collections in iRODS}
\usage{
iquery(
  query,
  limit = 100,
  offset = 0,
  type = c("general", "specific"),
  case_sensitive = TRUE,
  distinct = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{query}{GeneralQuery for searching the iCAT database.}

\item{limit}{Maximum number of rows to return. Defaults to 100.}

\item{offset}{Number of rows to skip for paging. Defaults to 0.}

\item{type}{Type of query: 'general' (the default) or 'specific'.}

\item{case_sensitive}{Whether the string matching in the query is case
sensitive. Defaults to \code{TRUE}.}

\item{distinct}{Whether only distinct rows should be listed. Defaults to
\code{TRUE}.}

\item{verbose}{Whether information should be printed about the HTTP request
and response. Defaults to \code{FALSE}.}
}
\value{
Invisibly, the HTTP response.
}
\description{
Use SQL-like expressions to query data objects and collections based on
different properties. Read more about queries by looking at the iCommands
equivalent \code{iquest} in the \href{https://docs.irods.org/master/icommands/user/#iquest}{iRODS Docs}.
}
\examples{
# demonstration server (requires Bash, Docker and Docker-compose)
irods_demo <- try(use_irods_demo())

if (!inherits(irods_demo, "try-error")) {

  # move to temporary directory and save old working directory
  old_dir <- getwd()
  tmp <- tempdir()
  setwd(tmp)

  # connect project to server
  create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home",
    overwrite = TRUE)

  # authentication
  iauth("rods", "rods")

  # search for objects by metadata
  iquery("SELECT COLL_NAME, DATA_NAME WHERE COLL_NAME LIKE '/tempZone/home/\%'")
  iquery("SELECT COLL_NAME, DATA_NAME WHERE META_DATA_ATTR_NAME LIKE 'foo\%'")

  stop_irods_demo()

  # back to previous directory
  setwd(old_dir)
}
}
\references{
https://docs.irods.org/master/icommands/user/#iquest
}
