% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods-http.R
\name{irods_http_call}
\alias{irods_http_call}
\title{Internal function to coordinate iRODS HTTP call}
\usage{
irods_http_call(endpoint, verb, args, verbose, error = TRUE)
}
\arguments{
\item{endpoint}{HTTP API end-point.}

\item{verb}{The HTTP verb to be used (e.g. PUT).}

\item{args}{The arguments to be set for the end-point.}

\item{verbose}{Verbosity of the HTTP request.}

\item{error}{Whether status codes larger than 400 should be translated to
errors.}
}
\value{
HTTP response
}
\description{
Internal function to coordinate iRODS HTTP call
}
\references{
https://github.com/irods/irods_client_http_api
}
\keyword{internal}
