% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods-rest.R
\name{irods_rest_call}
\alias{irods_rest_call}
\title{Internal function to coordinate iRODS REST call}
\usage{
irods_rest_call(endpoint, verb, args, verbose, object = NULL, error = TRUE)
}
\arguments{
\item{endpoint}{REST API end-point.}

\item{verb}{The HTTP verb to be used (e.g. PUT).}

\item{args}{The arguments to be set for the end-point.}

\item{verbose}{Verbosity of the HTTP request.}

\item{object}{Data to send along with HTTP request.}

\item{error}{Whether status codes larger than 400 should be translated to
errors.}
}
\value{
HTTP response
}
\description{
Internal function to coordinate iRODS REST call
}
\references{
https://github.com/irods/irods_demo/tree/main/irods_client_rest_cpp
}
\keyword{internal}
