% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods-auth.R
\name{is_connected_irods}
\alias{is_connected_irods}
\title{Predicate for iRODS Connectivity}
\usage{
is_connected_irods(...)
}
\arguments{
\item{...}{Currently not implemented.}
}
\value{
Boolean whether or not a connection to iRODS exists.
}
\description{
A predicate to check whether you are currently connected to an iRODS server.
}
\examples{
is_connected_irods()
}
