% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods-demo.R
\name{is_irods_demo_running}
\alias{is_irods_demo_running}
\title{Predicate check for iRODS demo server state}
\usage{
is_irods_demo_running()
}
\value{
boolean, TRUE if running, FALSE if stopped.
}
\description{
Predicate check for iRODS demo server state
}
\keyword{internal}
