% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev-helpers.R
\name{local_create_irods}
\alias{local_create_irods}
\title{Launch iRODS from alternative directory}
\usage{
local_create_irods(
  host = NULL,
  zone_path = NULL,
  dir = tempdir(),
  env = parent.frame()
)
}
\arguments{
\item{host}{Hostname of the iRODS server
(default: "http://localhost/irods-rest/0.9.3").}

\item{zone_path}{Zone path of the iRODS server.}

\item{dir}{The directory to use (default is temporary directory).}

\item{env}{Attach exit handlers to this environment. Defaults to the
parent frame (accessed through \code{parent.frame()}).}
}
\value{
Invisibly returns the original directory.
}
\description{
This function is useful during development as it prevents cluttering of the
package source files.
}
\examples{
# demonstration server (requires Bash, Docker and Docker-compose)
irods_demo <- try(use_irods_demo())

if (!inherits(irods_demo, "try-error")) {
  # launch iRODS from temporary directory
  rirods:::local_create_irods()
}
}
\keyword{internal}
