% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods-s3.R
\name{new_irods_df}
\alias{new_irods_df}
\title{iRODS Zone information class}
\usage{
new_irods_df(x = list())
}
\arguments{
\item{x}{list with iRODS Zone information.}
}
\value{
\code{irods_df} class object.
}
\description{
iRODS Zone information class
}
\keyword{internal}
