% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods-demo.R
\name{remove_docker_images}
\alias{remove_docker_images}
\title{Remove Docker images from system}
\usage{
remove_docker_images()
}
\value{
Invisible
}
\description{
Remove Docker images from system
}
\keyword{internal}
