% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{riadmin}
\alias{riadmin}
\title{The administration interface to the iRODS}
\usage{
riadmin(host = "http://localhost/irods-rest/0.9.2", token, action, target, ...)
}
\arguments{
\item{host}{Url of host}

\item{token}{Token for authentication}

\item{action}{The action: add, modify, or remove.}

\item{target}{The subject of the action: user, zone, resource, childtoresc,
childfromresc, token, group, rebalance, unusedAVUs, specificQuery.}

\item{...}{Addition arguments (arg1 up to arg7)}
}
\value{
Invisible on success (200) otherwise http status.
}
\description{
The administration interface to the iRODS
}
\examples{

# get token
tk <- get_token()

# add user bobby
riadmin(token = tk, action = "add", target = "user", arg2 = "bobby", arg3 = "rodsuser")

# check if bobby is added
rils(tk)

# remove user bobby
riadmin(token = tk, action = "remove", target = "user", arg2 = "bobby")

# check if bobby is removed
rils(tk)

}
