% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ils.R
\name{rils}
\alias{rils}
\title{Listing}
\usage{
rils(
  host = "http://localhost/irods-rest/0.9.2",
  path = "/tempZone/home",
  token,
  stat = FALSE,
  permissions = FALSE,
  metadata = FALSE,
  offset = 0,
  limit = 100
)
}
\arguments{
\item{host}{Url of host}

\item{path}{Directory to be listed.}

\item{token}{Token for authentication}

\item{stat}{Boolean flag to indicate stat information is desired}

\item{permissions}{Boolean flag to indicate access control information is desired}

\item{metadata}{Boolean flag to indicate metadata is desired}

\item{offset}{Number of records to skip for pagination}

\item{limit}{Number of records desired per page}
}
\value{
tibble with logical paths
}
\description{
Recursive listing of a collection, or stat, metadata, and access control
information for a given data object.
}
\examples{

# list home directory
rils(token = get_token())

}
