% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods-demo.R
\name{use_irods_demo}
\alias{use_irods_demo}
\alias{stop_irods_demo}
\alias{remove_docker_images}
\title{Run Docker iRODS demonstration service}
\usage{
use_irods_demo(user = character(), pass = character())

stop_irods_demo()

remove_docker_images()
}
\arguments{
\item{user}{Character vector for user name (defaults to "rods" admin)}

\item{pass}{Character vector for password (defaults to "rods" admin password)}
}
\value{
Invisible
}
\description{
Run an iRODS demonstration server as an Docker container instance.
The functions \code{stop_irods_demo()} and \code{remove_docker_images()} are used to
stop the containers, and subsequently remove the images.
}
\details{
These functions are untested on Windows and macOS and require:
\itemize{
\item \code{bash}
\item \code{docker}
\item \code{docker-compose}
}
}
\examples{

if (interactive()) {

  # launch docker irods_demo containers (and possibly download images) with
  # default credentials.
  use_irods_demo()

  # same but then with "alice" as user and "PASSword" as password
  use_irods_demo("alice", "PASSword")

  # stop containers
  stop_irods_demo()
}

}
\references{
https://github.com/irods/irods_demo
}
