% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{auth}
\alias{auth}
\title{Authentication service for the iRODS zone}
\usage{
auth(token = get_token())
}
\arguments{
\item{token}{Token for authentication}
}
\value{
String
}
\description{
Authentication service for the iRODS zone
}
\examples{

# authenticate
auth()

}
