% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-irods.R
\name{create_irods}
\alias{create_irods}
\title{Generate IRODS project file}
\usage{
create_irods(host, zone_path, path = ".", overwrite = FALSE)
}
\arguments{
\item{host}{URL of host.}

\item{zone_path}{Path to the zone of the iRODS server (e.g., "/tempZone/home").}

\item{path}{Path of project. Defaults to the top directory of the zone.}

\item{overwrite}{Overwrite existing iRODS project file. Defaults to \code{FALSE}.}
}
\value{
Invisibly, the path to the iRODS project file.
}
\description{
This will create an iRODS project file containing information about the
iRODS server. Once the file has been created, future sessions
connect again with the same iRODS server without further intervention.
}
\examples{
if(interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.2", "/tempZone/home")
}
}
