% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admin.R
\name{iadmin}
\alias{iadmin}
\title{The administration interface to the iRODS}
\usage{
iadmin(
  host = "http://localhost/irods-rest/0.9.2",
  action,
  target,
  ...,
  verbose = FALSE
)
}
\arguments{
\item{host}{Url of host}

\item{action}{The action: add, modify, or remove.}

\item{target}{The subject of the action: user, zone, resource, childtoresc,
childfromresc, token, group, rebalance, unusedAVUs, specificQuery.}

\item{...}{Addition arguments (arg1 up to arg7)}

\item{token}{Token for authentication}
}
\value{
Invisible on success (200) otherwise http status.
}
\description{
The administration interface to the iRODS
}
\examples{

# authenticate
auth()

# add user bobby
iadmin(action = "add", target = "user", arg2 = "bobby", arg3 = "rodsuser")

# check if bobby is added
ils()

# remove user bobby
iadmin(action = "remove", target = "user", arg2 = "bobby")

# check if bobby is removed
ils()

}
