% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{iauth}
\alias{iauth}
\title{Authentication service for the iRODS zone}
\usage{
iauth(user = NULL, password = NULL)
}
\arguments{
\item{user}{iRODS user name.}

\item{password}{iRODS password.}
}
\value{
Invisibly NULL
}
\description{
Authentication service for the iRODS zone
}
\examples{
if(interactive()) {
# authenticate
iauth()
}
}
