% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{iauth}
\alias{iauth}
\title{Authentication service for the iRODS zone}
\usage{
iauth(user = NULL, password = NULL, role = "rodsuser")
}
\arguments{
\item{user}{iRODS user name (prompts user for user name if not supplied).}

\item{password}{iRODS password (prompts user for password if not supplied).}

\item{role}{iRODS role of user (defaults to \code{"rodsuser"}).}
}
\value{
Invisibly \code{NULL}.
}
\description{
Provides an authentication service for the iRODS zone. Using the function
without arguments results in a prompt asking for the user name and password
thereby preventing hard-coding of sensitive information in scripts.
}
\examples{
if(interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home")
# authenticate
iauth()
}
}
