% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{iauth}
\alias{iauth}
\title{Authentication service for the iRODS zone}
\usage{
iauth(user = "rods", password = "rods")
}
\value{
String
}
\description{
Authentication service for the iRODS zone
}
\examples{

# authenticate
iauth() # i

}
