% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{icd}
\alias{icd}
\alias{ipwd}
\title{Get or set current working directory in iRODS}
\usage{
icd(dir)

ipwd()
}
\arguments{
\item{dir}{Collection to set as working directory.}
}
\value{
The current working directory (invisibly in the case of \code{cd()}).
}
\description{
\code{ipwd()} and \code{icd()} are the iRODS equivalents of \code{getwd()} and \code{setwd()}
respectively. For example, whereas \code{getwd()} returns the current working directory
in the local system, \code{ipwd()} returns the current working directory in iRODS.
}
\examples{
if(interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home")

# authenticate
iauth()

# default dir
icd(".")
ipwd()

# relative paths work as well
icd("/tempZone/home")
ipwd()

# go back on level lower
icd("..")
ipwd()

# absolute paths work as well
icd("/tempZone/home/rods")
ipwd()
}
}
