% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd.R
\name{icd}
\alias{icd}
\title{Retrieve an object from iRODS}
\usage{
icd(dir)
}
\arguments{
\item{dir}{Change the current directory to DIR.  The default DIR is the value
of the HOME shell variable.}
}
\value{
invisibly returns the path
}
\description{
Retrieve an object from iRODS
}
\examples{

# authenticate
auth()

# default dir
icd(".")
ils()

# some other dir
icd("/tempZone/home")
ils()

# go back on level lower
icd("..")
ils()

# relative paths work as well
icd("home/public")
ils()

}
