% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{icd}
\alias{icd}
\alias{ipwd}
\title{Retrieve an object from iRODS}
\usage{
icd(dir)

ipwd()
}
\arguments{
\item{dir}{Change the current directory to DIR.  The default DIR is the value
of the HOME shell variable.}
}
\value{
Visibly or invisibly returns the path.
}
\description{
Retrieve an object from iRODS
}
\examples{
if(interactive()) {
# authenticate
iauth()

# default dir
icd(".")
ipwd()


# some other dir
icd("/tempZone/home")
ipwd()

# go back on level lower
icd("..")
ipwd()

# relative paths work as well
icd("../home/public")
ipwd()
}
}
