% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{icd}
\alias{icd}
\alias{ipwd}
\title{Change/show current working directory}
\usage{
icd(dir)

ipwd()
}
\arguments{
\item{dir}{Change the current directory to DIR.  The default DIR is the value
of the HOME shell variable.}
}
\value{
Visibly or invisibly returns the path.
}
\description{
Changes iRODS the current working directory (collection). The functions
mimic behavior of unix \code{cd} and \code{pwd}.
}
\examples{
if(interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home")

# authenticate
iauth()

# default dir
icd(".")
ipwd()


# relative paths work as well
icd("/tempZone/home")
ipwd()

# go back on level lower
icd("..")
ipwd()

# absolute paths work as well
icd("/tempZone/home/rods")
ipwd()
}
}
