% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{icd}
\alias{icd}
\alias{ipwd}
\title{Get or Set Current Working Directory (Collection) in iRODS}
\usage{
icd(dir)

ipwd()
}
\arguments{
\item{dir}{Path to set as working collection.}
}
\value{
Invisibly the current collection before the change (same convention as
\code{setwd()}).
}
\description{
\code{ipwd()} and \code{icd()} are the iRODS equivalents of \code{\link[=getwd]{getwd()}} and \code{\link[=setwd]{setwd()}}
respectively. For example, whereas \code{getwd()} returns the current working directory
in the local system, \code{ipwd()} returns the current working collection in iRODS.
}
\examples{
\dontshow{if (is_irods_demo_running()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
is_irods_demo_running()
\dontshow{
.old_config_dir <- Sys.getenv("R_USER_CONFIG_DIR")
Sys.setenv("R_USER_CONFIG_DIR" = tempdir())
}
# connect project to server
\Sexpr[stage=build, results=rd]{paste0("create_irods(\"", rirods:::.irods_host, "\")")}

# authenticate
iauth("rods", "rods", "rodsadmin")

# default dir
icd(".")
ipwd()

# relative paths work as well
icd("/tempZone/home")
ipwd()

# go back on level lower
icd("..")
ipwd()

# absolute paths work as well
icd("/tempZone/home/rods")
ipwd()

# back home
icd("/tempZone/home")
\dontshow{
Sys.setenv("R_USER_CONFIG_DIR" = .old_config_dir)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=setwd]{setwd()}} and \code{\link[=getwd]{getwd()}} for R equivalents,
\code{\link[=ils]{ils()}} for listing collections and objects in iRODS.
}
