% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iget.R
\name{iget}
\alias{iget}
\title{Retrieve an object from iRODS}
\usage{
iget(
  host = "http://localhost/irods-rest/0.9.2",
  data,
  offset = 0,
  count = 1000
)
}
\arguments{
\item{host}{Url of host}

\item{data}{R object stored on iRODS server.}

\item{offset}{Offset in bytes into the data object (Defaults to 0)}

\item{count}{Maximum number of bytes to read or write}
}
\value{
R object
}
\description{
Retrieve an object from iRODS
}
\examples{

# authentication
auth()

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# set working directory to rods
icd("/tempZone/home/rods")

# store
iput(data = foo)

# check if file is stored
ils()

# retrieve in native R format
iget(data = "foo")

}
