% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/navigation.R
\name{ils}
\alias{ils}
\title{Listing iRODS data objects and collections}
\usage{
ils(
  logical_path = ".",
  stat = FALSE,
  permissions = FALSE,
  metadata = FALSE,
  offset = 0,
  limit = 100,
  message = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{logical_path}{Directory to be listed.}

\item{stat}{Boolean flag to indicate stat information is desired.}

\item{permissions}{Boolean flag to indicate access control information is
desired.}

\item{metadata}{Boolean flag to indicate metadata is desired.}

\item{offset}{Number of records to skip for pagination.}

\item{limit}{Number of records desired per page.}

\item{message}{In case the collection is empty a message saying so is
returned.}

\item{verbose}{Show information about the http request and response.}
}
\value{
tibble with logical paths
}
\description{
Recursive listing of a collection, or stat, metadata, and access control
information for a given data object.
}
\examples{
if(interactive()) {
# authenticate
iauth()

# list home directory
ils()
}
}
