% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{imeta}
\alias{imeta}
\title{Add or remove metadata}
\usage{
imeta(
  logical_path,
  entity_type = c("data_object", "collection", "user"),
  operations = list(),
  verbose = FALSE
)
}
\arguments{
\item{logical_path}{Path to the data object or collection (or name of the user).}

\item{entity_type}{Type of item to add metadata to or remove it from.
Options are 'data_object', 'collection' and 'user'.}

\item{operations}{List of named lists of vectors representing operations.
The valid components of each of these lists or vectors are:
\itemize{
\item \code{operation}, with values 'add' or 'remove', depending on whether the AVU
should be added to or removed from the metadata of the item.
\item \code{attribute}, with the name of the AVU.
\item \code{value}, with the value of the AVU.
\item \code{units}, with the unit of the AVU.
}}

\item{verbose}{Whether information should be printed about the HTTP request and response.}
}
\value{
Invisibly, the HTTP response.
}
\description{
In iRODS, metadata is stored as attribute-value-units triples (AVUs), consisting
of an attribute name, an attribute value and an optional unit.
This function allows to chain several operations ('add' or 'remove') linked to
specific AVUs.
}
\examples{
if(interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home")

# authentication
iauth()

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# store
iput(foo, "foo.rds")

# check if file is stored
ils()

# add some metadata
imeta(
 "foo.rds",
 "data_object",
 operations = list(
    operation = "add",
    attribute = "foo",
    value = "bar",
    units = "baz"
    )
)

imeta(
 "foo.rds",
 "data_object",
 operations = list(
   list(operation = "add", attribute = "foo2", value = "bar2"),
   list(operation = "add", attribute = "foo3", value = "bar3")
 )
)

# check if file is stored with associated metadata
ils(metadata = TRUE)
}
}
