% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata.R
\name{imeta}
\alias{imeta}
\alias{iquery}
\title{Describe your iRODS objects}
\usage{
imeta(
  logical_path,
  entity_type,
  operations = list(operation = "add", attribute = NULL, value = NULL, units = NULL),
  path = ".",
  verbose = FALSE
)

iquery(
  query,
  limit = 100,
  offset = 0,
  type = "general",
  case_sensitive = TRUE,
  distinct = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{logical_path}{object, collection or user}

\item{entity_type}{Type (object, collection or user)}

\item{operations}{List which contains the following named elements;
\code{operation}, which adds (\code{"add"}) or removes (\code{"remove"}) the object, and
the metadata triplet attribute, value, units.}

\item{path}{Path to object (defaults to \code{"."}).}

\item{verbose}{Show information about the http request and response.
query}

\item{query}{GeneralQuery for searching the iCAT database.}

\item{limit}{The max number of rows to return (defaults to 100)}

\item{offset}{Number of rows to skip for paging (defaults to 0).}

\item{type}{Either 'general' or 'specific' (defaults to 'general').}

\item{case_sensitive}{Affects string matching (defaults to TRUE).}

\item{distinct}{Only list distinct rows (defaults to TRUE).}
}
\value{
Invisibly returns the response.
}
\description{
Metadata attribute-value-units triples (AVUs) consist of an Attribute-Name,
Attribute-Value, and an optional Attribute-Units. They can be added
via the 'add' command ,or removed with 'remove', and then queried to find
matching objects.
}
\examples{
if(interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home")

# authentication
iauth()

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# store
iput(foo, "foo.rds")

# check if file is stored
ils()

# add some metadata
imeta(
 "foo.rds",
 "data_object",
 operations =
  list(operation = "add", attribute = "foo", value = "bar", units = "baz")
)

# check if file is stored with associated metadata
ils(metadata = TRUE)

# search for objects by metadata
iquery("SELECT COLL_NAME, DATA_NAME WHERE COLL_NAME LIKE '/tempZone/home/\%'")
}
}
