% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{imkdir}
\alias{imkdir}
\title{Create a new collection in iRODS}
\usage{
imkdir(
  logical_path,
  collection = TRUE,
  create_parent_collections = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{logical_path}{Path to the collection to create, relative to the current
working directory (see \code{ipwd()}).}

\item{collection}{Whether a collection is being created. Defaults to \code{TRUE}.}

\item{create_parent_collections}{Whether parent collections should be created
when necessary.}

\item{verbose}{Whether information about the HTTP request and response should be printed.}
}
\value{
Invisibly the HTTP request.
}
\description{
This is the equivalent to \code{dir.create()}, but creating a collection in iRODS
instead of a local directory.
}
\examples{
if (interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home")

# authentication
iauth()

ils()
imkdir("new_collection")
ils()
icd("new_collection")
}
}
