% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-objects.R
\name{iput}
\alias{iput}
\alias{isaveRDS}
\title{Save files and objects in iRODS}
\usage{
iput(
  local_path,
  logical_path = basename(local_path),
  offset = 0,
  count = 2000L,
  truncate = "true",
  verbose = FALSE,
  overwrite = FALSE
)

isaveRDS(
  x,
  logical_path,
  offset = 0,
  count = 2000L,
  truncate = "true",
  verbose = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{local_path}{Local path of file to be sent to iRODS.}

\item{logical_path}{Destination path in iRODS.
By default, the basename of \code{local_path} is used and the file is
stored in the current working directory (see \code{ipwd()}).}

\item{offset}{Offset in bytes into the data object. Defaults to 0.}

\item{count}{Maximum number of bytes to write. Defaults to 2000.}

\item{truncate}{Whether to truncate the object when opening it. Defaults to \code{TRUE}.}

\item{verbose}{Whether to print information about the HTTP request and response.}

\item{overwrite}{Whether the file in iRODS should be overwritten if it exists.
Defaults to \code{FALSE}.}

\item{x}{R object to save in iRODS.}
}
\value{
(Invisibly) the HTTP response.
}
\description{
Store an object or file into iRODS. \code{iput()} should be used to transfer
a file from the local storage to iRODS; \code{isaveRDS()} saves an R object from the
current environment in iRODS in RDS format (see \code{saveRDS()}).
}
\examples{
if (interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home", TRUE)

# authenticate
iauth()

# save the iris dataset as csv and send the file to iRODS
library(readr)
write_csv(iris, "iris.csv")
iput("iris.csv", overwrite = TRUE)

# save with a different name
iput("iris.csv", "irids_in_irods.csv", overwrite = TRUE)
ils()

# send an R object to iRODS in RDS format
isaveRDS(iris, "irids_in_rds.rds", overwrite = TRUE)
}
}
