% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iput.R
\name{iput}
\alias{iput}
\title{Store an object into iRODS}
\usage{
iput(
  host = "http://localhost/irods-rest/0.9.2",
  path = "/tempZone/home",
  file,
  offset = 0
)
}
\arguments{
\item{host}{Url of host}

\item{file}{R object stored on iRODS server.}

\item{offset}{Offset in bytes into the data object (Defaults to 0)}
}
\value{
http response message
}
\description{
Store an object into iRODS
}
\examples{

# authenticate
auth()

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# store
iput(file = foo, path = "/tempZone/home/rods")

}
