% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-objects.R
\name{iput}
\alias{iput}
\alias{iget}
\title{Working with data objects}
\usage{
iput(x, path = ".", offset = 0, verbose = FALSE)

iget(
  host = "http://localhost/irods-rest/0.9.2",
  data,
  offset = 0,
  count = 1000,
  verbose = FALSE
)
}
\arguments{
\item{x}{R object stored on iRODS server.}

\item{offset}{Offset in bytes into the data object (Defaults to 0).}

\item{verbose}{Show information about the http request and response.}

\item{host}{Url of host}

\item{count}{Maximum number of bytes to read or write.}
}
\value{
R object
}
\description{
Working with data objects
}
\examples{

# authentication
iauth("bobby", "passWORD")

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# store
iput(foo)

# check if file is stored
ils()

# retrieve in native R format
iget("foo")

}
