% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-objects.R
\name{iput}
\alias{iput}
\alias{iget}
\title{Working with data objects}
\usage{
iput(
  x,
  path = ".",
  offset = 0,
  truncate = TRUE,
  verbose = FALSE,
  overwrite = FALSE
)

iget(x, path = x, offset = 0, count = 1000, verbose = FALSE, overwrite = FALSE)
}
\arguments{
\item{x}{R object stored on iRODS server.}

\item{path}{Destination path.}

\item{offset}{Offset in bytes into the data object (Defaults to FALSE).}

\item{truncate}{Truncates the object on open (defaults to TRUE).}

\item{verbose}{Show information about the http request and response.}

\item{overwrite}{Overwrite irods object or local file (defaults to FALSE).}

\item{count}{Maximum number of bytes to read or write.}
}
\value{
R object
}
\description{
Working with data objects
}
\examples{
if(interactive()) {
# authentication
iauth()

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# store
iput(foo)

# check if file is stored
ils()

# retrieve in native R format
iget("foo")
}
}
