% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irm.R
\name{irm}
\alias{irm}
\title{Delete an object from iRODS}
\usage{
irm(
  host = "http://localhost/irods-rest/0.9.2",
  file = "/tempZone/home",
  trash = TRUE,
  recursive = FALSE,
  unregister = FALSE
)
}
\arguments{
\item{host}{Url of host}

\item{file}{R object stored on iRODS server.}

\item{trash}{Send to trash or delete permanently (default = TRUE).}

\item{recursive}{Recursively delete contents of a collection
(default = FALSE).}

\item{unregister}{Unregister data objects instead of deleting them
(default = FALSE).}
}
\description{
Delete an object from iRODS
}
