% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{irm}
\alias{irm}
\title{Remove data objects or collections in iRODS.}
\usage{
irm(logical_path, force = TRUE, recursive = FALSE, verbose = FALSE)
}
\arguments{
\item{logical_path}{Path to the data object or collection to remove.}

\item{force}{Whether the data object or collection should be deleted permanently.
If \code{FALSE}, it is sent to the trash collection.}

\item{recursive}{If a collection is provided, whether its contents should also be
removed. If a collection is not empty and \code{recursive} is \code{FALSE}, it cannot
be deleted.}

\item{verbose}{Whether information should be printed about the HTTP request and response.}
}
\value{
Invisibly the HTTP call.
}
\description{
This is the equivalent of \code{file.remove()}, but applied to an item inside iRODS.
}
\examples{
if (interactive()) {
# connect project to server
create_irods("http://localhost/irods-rest/0.9.3", "/tempZone/home")

# authentication
iauth()

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# store
iput(foo, "foo.rds")

# check if file is stored
ils()

# delete object
irm("foo.rds", force = TRUE)
iquery("SELECT COLL_NAME, DATA_NAME WHERE DATA_NAME LIKE 'foo\%'")
}
}
