% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{irm}
\alias{irm}
\title{Delete an object from iRODS}
\usage{
irm(
  host = "http://localhost/irods-rest/0.9.2",
  data,
  trash = TRUE,
  recursive = FALSE,
  unregister = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{host}{Url of host}

\item{data}{R object stored on iRODS server.}

\item{trash}{Send to trash or delete permanently (default = TRUE).}

\item{recursive}{Recursively delete contents of a collection
(default = FALSE).}

\item{unregister}{Unregister data objects instead of deleting them
(default = FALSE).}

\item{verbose}{Show information about the http request and response.}
}
\value{
Invisible object to be removed
}
\description{
Delete an object from iRODS
}
\examples{

# authentication
auth()

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# set working directory to rods
icd("/tempZone/home/rods")

# store
iput(data = foo)

# check if file is stored
ils()

# retrieve in native R format
iget(data = "foo")

# delete object

irm(data = "foo", trash = FALSE)

# check if file is delete
ils()
}
