% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{irm}
\alias{irm}
\title{Delete an object from iRODS}
\usage{
irm(x, trash = TRUE, recursive = FALSE, unregister = FALSE, verbose = FALSE)
}
\arguments{
\item{x}{R object stored on iRODS server.}

\item{trash}{Send to trash or delete permanently (default = TRUE).}

\item{recursive}{Recursively delete contents of a collection
(default = FALSE).}

\item{unregister}{Unregister data objects instead of deleting them
(default = FALSE).}

\item{verbose}{Show information about the http request and response.}
}
\value{
Invisible object to be removed
}
\description{
Delete an object from iRODS
}
\examples{
if(interactive()) {
# authentication
iauth()

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# store
iput(foo)

# check if file is stored
ils()

# delete object
irm("foo", trash = FALSE)

# check if file is delete
ils()
}
}
