% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collections.R
\name{irm}
\alias{irm}
\alias{imkdir}
\title{Create/remove objects in iRODS}
\usage{
irm(x, trash = TRUE, recursive = FALSE, unregister = FALSE, verbose = FALSE)

imkdir(
  x,
  collection = TRUE,
  create_parent_collections = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{x}{R object stored on iRODS server.}

\item{trash}{Send to trash or delete permanently (default = TRUE).}

\item{recursive}{Recursively delete contents of a collection
(default = FALSE).}

\item{unregister}{Unregister data objects instead of deleting them
(default = FALSE).}

\item{verbose}{Show information about the http request and response.}

\item{collection}{Indicates that a collection is being created (defaults to
TRUE).}

\item{create_parent_collections}{Indicates that parent collections of the
destination collection should be created if necessary. Only applicable when
collection = TRUE is used. Defaults to FALSE.}
}
\value{
Invisible object to be removed
}
\description{
These functions provide an interface to remove or create objects or
collections from the iRODS namespace. By default the objects/collections are
moved to the trash collection when removed.
}
\examples{
if(interactive()) {
# authentication
iauth()

# some data
foo <- data.frame(x = c(1, 8, 9), y = c("x", "y", "z"))

# store
iput(foo)

# check if file is stored
ils()

# delete object
irm("foo", trash = FALSE)

# create collection
imkdir("a")

# check if file is delete
ils()
}
}
