% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods-demo.R
\name{is_irods_demo_running}
\alias{is_irods_demo_running}
\title{Predicate for iRODS Demonstration Service State}
\usage{
is_irods_demo_running(...)
}
\arguments{
\item{...}{Currently not implemented.}
}
\value{
Boolean whether or not connected to iRODS
}
\description{
A predicate to check whether you are running iRODS docker demo containers.
}
\examples{
is_irods_demo_running()
}
