% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irods-demo.R
\name{local_create_irods}
\alias{local_create_irods}
\title{Launch iRODS from Alternative Directory}
\usage{
local_create_irods(host = NULL, dir = tempdir(), env = parent.frame())
}
\arguments{
\item{host}{Hostname of the iRODS server. Defaults to
""http://localhost:9001/irods-http-api/0.1.0".}

\item{dir}{The directory to use. Default is a temporary directory.}

\item{env}{Attach exit handlers to this environment. Defaults to the
parent frame (accessed through \code{\link[=parent.frame]{parent.frame()}}).}
}
\value{
Invisibly returns the original directory.
}
\description{
This function is useful during development as it prevents cluttering of the
package source files.
}
\keyword{internal}
