% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create-irods.R
\name{local_create_irods}
\alias{local_create_irods}
\title{Launch iRODS from Alternative Directory}
\usage{
local_create_irods(
  host = NULL,
  zone_path = NULL,
  dir = tempdir(),
  env = parent.frame()
)
}
\arguments{
\item{host}{Hostname of the iRODS server. Defaults to
"http://localhost/irods-rest/0.9.3".}

\item{zone_path}{Zone path of the iRODS server. Defaults to "/tempZone/home".}

\item{dir}{The directory to use. Default is a temporary directory.}

\item{env}{Attach exit handlers to this environment. Defaults to the
parent frame (accessed through \code{\link[=parent.frame]{parent.frame()}}).}
}
\value{
Invisibly returns the original directory.
}
\description{
This function is useful during development as it prevents cluttering of the
package source files.
}
