% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{local_create_irods}
\alias{local_create_irods}
\title{Launch iRODS from alternative directory}
\usage{
local_create_irods(
  host = NULL,
  zone_path = NULL,
  dir = tempdir(),
  env = parent.frame()
)
}
\arguments{
\item{host}{Hostname of the iRODS server
(default: "http://localhost/irods-rest/0.9.3").}

\item{zone_path}{Zone path of the iRODS server.}

\item{dir}{The directory to use (default is temporary directory).}

\item{env}{Attach exit handlers to this environment. Defaults to the
parent frame (accessed through parent.frame()).}
}
\value{
Invisibly returns the original directory.
}
\description{
This function is useful during development as it prevents cluttering of the
package source files.
}
\examples{
if (interactive()) {
  # launch iRODS from temporary directory
  local_create_irods()
}
}
