% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test-helpers.R
\name{remove_mock_files}
\alias{remove_mock_files}
\title{Remove http snapshots or mockfiles}
\usage{
remove_mock_files()
}
\value{
Invisibly the mock file paths.
}
\description{
Remove http snapshots or mockfiles
}
