% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rirods-package.R
\docType{package}
\name{rirods-package}
\alias{rirods}
\alias{rirods-package}
\title{rirods: R Client for iRODS}
\description{
The Integrated Rule-Oriented Data System 'iRODS' (\url{https://irods.org/}) open sourced data management software suite offer solutions which encompasses the whole data life cycle from data generation, storage and recycling. The loosely constructed and highly configurable architecture of iRODS frees the user from strict formatting constrains and single-vendor solutions. This package provides an R interface to iRODS' REST API, allowing you to manage your data from your own machine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/irods/irods_client_library_rirods}
  \item Report bugs at \url{https://github.com/irods/irods_client_library_rirods/issues}
}

}
\author{
\strong{Maintainer}: Martin Schobben \email{schobbenmartin@gmail.com} (\href{https://orcid.org/0000-0001-8560-0037}{ORCID})

Other contributors:
\itemize{
  \item Terrell Russell [contributor]
  \item Christine Staiger [contributor]
  \item Ton Smeele [contributor]
  \item Alan King [contributor]
}

}
\keyword{internal}
